will update with GET /api endpoint (ticket 3) with test suite - also renamed controllers and models for greater clarity moving forward
