+++ controllers/endpoints-controller.js
@@ -1,0 +1,5 @@
+const endpoints = require("../endpoints.json");
+
+exports.getEndpoints = (req, res) => {
+    res.status(200).send({ endpoints });
+}
